+++ packages/pwa/app/components/_app/index.jsx
@@ -9,6 +9,7 @@ import React, {useState, useEffect} from 'react'
 import PropTypes from 'prop-types'
 import {useHistory, useLocation} from 'react-router-dom'
 import {getAssetUrl} from 'pwa-kit-react-sdk/ssr/universal/utils'
+import {getAppOrigin} from 'pwa-kit-react-sdk/utils/url'
 
 // Chakra
 import {Box, useDisclosure, useStyleConfig} from '@chakra-ui/react'
@@ -39,9 +40,9 @@ import {defineMessages, IntlProvider} from 'react-intl'
 
 // Others
 import {watchOnlineStatus, flatten} from '../../utils/utils'
-import {homeUrlBuilder} from '../../utils/url'
+import {homeUrlBuilder, getUrlWithLocale} from '../../utils/url'
 import {getLocaleConfig, getPreferredCurrency} from '../../utils/locale'
-import {DEFAULT_CURRENCY, HOME_HREF} from '../../constants'
+import {DEFAULT_CURRENCY, HOME_HREF, SUPPORTED_LOCALES} from '../../constants'
 
 import Seo from '../seo'
 import useWishlist from '../../hooks/use-wishlist'
@@ -66,6 +67,8 @@ export const defaultLocaleMessages = defineMessages({
 
 const App = (props) => {
     const {children, targetLocale, defaultLocale, messages, categories: allCategories = {}} = props
+
+    const appOrigin = getAppOrigin()
 
     const history = useHistory()
     const location = useLocation()
@@ -171,6 +174,25 @@ const App = (props) => {
                                 href={getAssetUrl('static/img/global/apple-touch-icon.png')}
                             />
                             <link rel="manifest" href={getAssetUrl('static/manifest.json')} />
+
+                            {/* Urls for all localized versions of this page (including current page)
+                            For more details on hrefLang, see https://developers.google.com/search/docs/advanced/crawling/localized-versions */}
+                            {SUPPORTED_LOCALES.map((locale) => (
+                                <link
+                                    rel="alternate"
+                                    hrefLang={locale.id.toLowerCase()}
+                                    href={`${appOrigin}${getUrlWithLocale(locale.id, {location})}`}
+                                    key={locale.id}
+                                />
+                            ))}
+                            {/* A general locale as fallback. For example: "en" if default locale is "en-GB" */}
+                            <link
+                                rel="alternate"
+                                hrefLang={defaultLocale.slice(0, 2)}
+                                href={`${appOrigin}${getUrlWithLocale(defaultLocale, {location})}`}
+                            />
+                            {/* A wider fallback for user locales that the app does not support */}
+                            <link rel="alternate" hrefLang="x-default" href={`${appOrigin}/`} />
                         </Seo>
 
                         <ScrollToTop />
@@ -284,9 +306,9 @@ App.getProps = async ({api}) => {
         const message =
             rootCategory.title === 'Unsupported Locale'
                 ? `
-                
+
 🚫 This page isn’t working.
-It looks like the locale ‘${rootCategory.locale}’ hasn’t been set up, yet. 
+It looks like the locale ‘${rootCategory.locale}’ hasn’t been set up, yet.
 You can either follow this doc, https://sfdc.co/B4Z1m to enable it in business manager or define a different locale with the instructions for Localization in the README file.
 `
                 : rootCategory.detail
@@ -309,7 +331,6 @@ App.propTypes = {
     children: PropTypes.node,
     targetLocale: PropTypes.string,
     defaultLocale: PropTypes.string,
-    location: PropTypes.object,
     messages: PropTypes.object,
     categories: PropTypes.object
 }